+++ app/entities/offer/models_offer.py
