+++ app/routers/offer_routers.py
@@ -4,11 +4,11 @@ from fastapi import Depends, APIRouter, HTTPException, status
 from sqlalchemy.orm import joinedload, selectinload
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from entities.offer.offer_model import OfferLinkModel
+from entities.offer.models_offer import OfferLinkModel
 from entities.service_name.models_service_name import ServiceNameModel
 from routers.consts import RouteSlug
 from models.database import get_session
-from entities.offer.offer_model import OfferLinkInSchema
+from entities.offer.models_offer import OfferLinkInSchema
 from schemas.offer_schema import (
     OfferFullResponseSchema,
     OfferLinkOptionalSchema,
