+++ app/services/service_service.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from entities.offer.offer_model import OfferLinkModel
+from entities.offer.models_offer import OfferLinkModel
 from services.base_service import AbstractService
